requirements.txt.txt
streamlit==1.31.0
ultralytics==8.1.0
numpy
pandas
Pillow
opencv-python-headless